+++ cookbooks/arcgis-enterprise/attributes/portal.rb
@@ -58,6 +58,9 @@ default['arcgis']['portal'].tap do |portal|
     portal['keystore_password'] = ENV['ARCGIS_PORTAL_KEYSTORE_PASSWORD']
   end
   portal['cert_alias'] = node['arcgis']['portal']['domain_name']
+  portal['root_cert'] = ''
+  portal['root_cert_alias'] = ''
+  portal['tomcat_java_opts'] = ''
   portal['configure_autostart'] = true
   portal['install_system_requirements'] = true
   unless node['arcgis']['portal']['authorization_file'].nil?
@@ -81,11 +84,26 @@ default['arcgis']['portal'].tap do |portal|
     portal['lp-setup'] = node['arcgis']['server']['setup']
     portal['install_dir'] = ENV['ProgramW6432'] + '\\ArcGIS\\Portal'
     portal['data_dir'] = 'C:\\arcgisportal'
-    portal['local_content_dir'] = node['arcgis']['portal']['data_dir'] + '\\content'
-    portal['content_dir'] = node['arcgis']['portal']['local_content_dir']
-    portal['log_dir'] = node['arcgis']['portal']['data_dir'] + '\\logs'
+
+    if node['arcgis']['portal']['data_dir'].nil?
+      portal['local_content_dir'] = portal['data_dir'] + '\\content'
+      portal['log_dir'] = portal['data_dir'] + '\\logs'
+    else
+      portal['local_content_dir'] = node['arcgis']['portal']['data_dir'] + '\\content'
+      portal['log_dir'] = node['arcgis']['portal']['data_dir'] + '\\logs'
+    end
+
+    if node['arcgis']['portal']['local_content_dir'].nil?
+      portal['content_dir'] = portal['local_content_dir']
+    else
+      portal['content_dir'] = node['arcgis']['portal']['local_content_dir']
+    end
 
     case node['arcgis']['version']
+    when '10.6'
+      portal['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
+                                            'Portal_for_ArcGIS_Windows_106_161831.exe')
+      portal['product_code'] = '{FFE4808A-1AD2-41A6-B5AD-2BA312BE6AAA}'
     when '10.5.1'
       portal['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
                                             'Portal_for_ArcGIS_Windows_1051_156365.exe')
@@ -124,6 +142,9 @@ default['arcgis']['portal'].tap do |portal|
     portal['lp-setup'] = node['arcgis']['server']['setup']
 
     case node['arcgis']['version']
+    when '10.6'
+      portal['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
+                                            'Portal_for_ArcGIS_Linux_106_161809.tar.gz')
     when '10.5.1'
       portal['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
                                             'Portal_for_ArcGIS_Linux_1051_156440.tar.gz')
@@ -152,4 +173,6 @@ default['arcgis']['portal'].tap do |portal|
   portal['content_store_provider'] = 'FileSystem'
   portal['content_store_connection_string'] = node['arcgis']['portal']['content_dir']
   portal['object_store'] = nil
+  portal['upgrade_backup'] = true
+  portal['upgrade_rollback'] = true
 end