+++ cookbooks/arcgis-enterprise/providers/portal.rb
@@ -87,6 +87,8 @@ action :unpack do
 
     FileUtils.chown_R @new_resource.run_as_user, nil, repo
   end
+
+  new_resource.updated_by_last_action(true)
 end
 
 action :install do
@@ -310,6 +312,11 @@ action :create_site do
 
   portal_admin_client.wait_until_available
 
+  # Complete portal upgrade if the upgrade API is available (starting from ArcGIS Enterprise 10.6)
+  if portal_admin_client.complete_upgrade(@new_resource.upgrade_backup, @new_resource.upgrade_rollback)
+    portal_admin_client.post_upgrade()
+  end
+
   if portal_admin_client.site_exist?
     Chef::Log.info('Portal site already exists.')
   else
@@ -349,6 +356,12 @@ action :create_site do
                                           @new_resource.log_dir,
                                           @new_resource.max_log_file_age)
 
+    unless (@new_resource.root_cert.empty? || @new_resource.root_cert_alias.empty?)
+      portal_admin_client.add_root_cert(@new_resource.root_cert, @new_resource.root_cert_alias, 'false')
+    end
+
+    sleep(60.0) # wait for portal restart
+
     new_resource.updated_by_last_action(true)
   end
 end
@@ -360,6 +373,9 @@ action :join_site do
     @new_resource.password)
 
   portal_admin_client.wait_until_available
+
+  # Complete portal upgrade if the upgrade API is available (starting from ArcGIS Enterprise 10.6)
+  portal_admin_client.complete_upgrade(@new_resource.upgrade_backup, @new_resource.upgrade_rollback)
 
   if portal_admin_client.site_exist?
     Chef::Log.info('Portal site already exists.')
@@ -571,6 +587,17 @@ action :start do
       new_resource.updated_by_last_action(true)
     end
   else
+    tomcat_java_opts = node['arcgis']['portal']['tomcat_java_opts']
+    unless (tomcat_java_opts.empty?)
+      replace = "indexserver"
+      pattern = "#{replace} #{tomcat_java_opts}"
+      catalina_file = ::File.join(node['arcgis']['portal']['install_dir'], node['arcgis']['portal']['install_subdir'], '/framework/runtime/tomcat/bin/catalina.sh')
+      sed_cmd = "sed -i \"s/#{replace}/#{pattern}/g\" #{catalina_file}"
+      sed_cmd = Mixlib::ShellOut.new(sed_cmd, {:timeout => 30})
+      sed_cmd.run_command
+      sed_cmd.error!
+    end
+    
     if node['arcgis']['portal']['configure_autostart']
       service 'arcgisportal' do
         supports :status => true, :restart => true, :reload => true