+++ cookbooks/arcgis-enterprise/recipes/enterprise_uninstalled.rb
@@ -19,9 +19,11 @@
 
 arcgis_enterprise_webadaptor 'Uninstall Web Adaptor for Portal' do
   install_dir node['arcgis']['web_adaptor']['install_dir']
-  product_code node['arcgis']['web_adaptor']['product_code2']
   instance_name node['arcgis']['portal']['wa_name']
   if node['platform'] == 'windows'
+    product_code Utils.wa_product_code(node['arcgis']['portal']['wa_name'],
+                                      [node['arcgis']['web_adaptor']['product_code'],
+                                       node['arcgis']['web_adaptor']['product_code2']])
     only_if { 
       !Utils.wa_product_code(node['arcgis']['portal']['wa_name'],
                             [node['arcgis']['web_adaptor']['product_code'],
@@ -38,9 +40,11 @@ end
 
 arcgis_enterprise_webadaptor 'Uninstall Web Adaptor for Server' do
   install_dir node['arcgis']['web_adaptor']['install_dir']
-  product_code node['arcgis']['web_adaptor']['product_code']
   instance_name node['arcgis']['server']['wa_name']
   if node['platform'] == 'windows'
+    product_code Utils.wa_product_code(node['arcgis']['server']['wa_name'],
+                                      [node['arcgis']['web_adaptor']['product_code'],
+                                       node['arcgis']['web_adaptor']['product_code2']])
     only_if {
       !Utils.wa_product_code(node['arcgis']['server']['wa_name'],
                             [node['arcgis']['web_adaptor']['product_code'],
@@ -59,15 +63,12 @@ arcgis_enterprise_portal 'Uninstall Portal for ArcGIS' do
   install_dir node['arcgis']['portal']['install_dir']
   product_code node['arcgis']['portal']['product_code']
   if node['platform'] == 'windows'
-    only_if {
-      Utils.product_installed?(node['arcgis']['portal']['product_code'])
-    }
+    only_if { Utils.product_installed?(node['arcgis']['portal']['product_code']) }
   else
-    only_if {
-      ::File.exist?(::File.join(node['arcgis']['portal']['install_dir'],
-                                node['arcgis']['portal']['install_subdir'],
-                                'startportal.sh'))
-    }
+    only_if { EsriProperties.product_installed?(node['arcgis']['run_as_user'],
+                                                node['hostname'],
+                                                node['arcgis']['version'],
+                                                :ArcGISPortal) }
   end
   action :uninstall
 end
@@ -76,15 +77,12 @@ arcgis_enterprise_datastore 'Uninstall ArcGIS DataStore' do
   install_dir node['arcgis']['data_store']['install_dir']
   product_code node['arcgis']['data_store']['product_code']
   if node['platform'] == 'windows'
-    only_if {
-      Utils.product_installed?(node['arcgis']['data_store']['product_code'])
-    }
+    only_if { Utils.product_installed?(node['arcgis']['data_store']['product_code']) }
   else
-    only_if {
-      ::File.exist?(::File.join(node['arcgis']['data_store']['install_dir'],
-                                node['arcgis']['data_store']['install_subdir'],
-                                'startdatastore.sh'))
-    }
+    only_if { EsriProperties.product_installed?(node['arcgis']['run_as_user'],
+                                                node['hostname'],
+                                                node['arcgis']['version'],
+                                                :ArcGISDataStore) }
   end
   action :uninstall
 end
@@ -93,15 +91,12 @@ arcgis_enterprise_server 'Uninstall ArcGIS Server' do
   install_dir node['arcgis']['server']['install_dir']
   product_code node['arcgis']['server']['product_code']
   if node['platform'] == 'windows'
-    only_if {
-      Utils.product_installed?(node['arcgis']['server']['product_code'])
-    }
+    only_if { Utils.product_installed?(node['arcgis']['server']['product_code']) }
   else
-    only_if {
-      ::File.exist?(::File.join(node['arcgis']['server']['install_dir'],
-                                node['arcgis']['server']['install_subdir'],
-                                'startserver.sh'))
-    }
+    only_if { EsriProperties.product_installed?(node['arcgis']['run_as_user'],
+                                                node['hostname'],
+                                                node['arcgis']['version'],
+                                                :ArcGISServer) }
   end
   action [:stop, :uninstall]
 end