+++ cookbooks/arcgis-geoevent/metadata.rb
@@ -4,7 +4,8 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Installs and configures ArcGIS GeoEvent Server'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '3.1.0'
+version          '3.2.0'
+chef_version     '>= 12.6', '< 14.0' if defined? chef_version
 
 depends          'arcgis-enterprise'