+++ cookbooks/arcgis-enterprise/CHANGELOG.md
@@ -3,6 +3,13 @@ arcgis-enterprise cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-enterprise cookbook.
 
+3.2.0
+-----
+- Support for ArcGIS Enterprise 10.6.
+- Support for Chef-Client 13.
+- Setting maximum heap size of ArcSOC processes.
+- Support upgrade of ArcGIS Enterprise on Linux.
+
 3.1.0
 -----
 - Added support for ArcGIS Enterprise 10.5.1.