+++ cookbooks/arcgis-geoevent/CHANGELOG.md
@@ -1,11 +1,16 @@
 arcgis-geoevent cookbook CHANGELOG
 ================
 
-This file is used to list changes made in each version of the arcgis-desktop cookbook.
+This file is used to list changes made in each version of the arcgis-geoevent cookbook.
+
+3.2.0
+-----
+- Added support for ArcGIS Enterprise 10.6
+- Support for Chef-Client 13.
 
 3.1.0
 -----
-- Added support for ArcGIS 10.5.1
+- Added support for ArcGIS Enterprise 10.5.1
 
 3.0.0
 -----